+++ renderer/src/camera.rs
@@ -28,7 +28,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
+    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
         Ray {
             origin: self.origin,
             dir: self.lower_left_corner + (self.horizontal * u) + (self.vertical * v) - self.origin,
